Lazy-load walk-in routes and extract team card models
