+++ apps/web/src/features/walk-in/TeamCardWalkInPage.tsx
@@ -1,12 +1,3 @@
-import type {
-	DeveloperDetails,
-	DeveloperFeatureUsage,
-	DeveloperProject,
-	DeveloperSession,
-	DimensionAnalysisDataPoint,
-	WrappedSourceSplit,
-	WrappedV1,
-} from "@rudel/api-routes";
 import { useDialKit } from "dialkit";
 import { ChevronLeft, ChevronRight, Clipboard, Download, Share2 } from "lucide-react";
 import {
@@ -32,6 +23,13 @@ import {
 import {
 	type WalkInOnboardingMetrics,
 } from "@/features/walk-in/walk-in-onboarding-types";
+import {
+	buildResolvedTeamCardRow,
+	buildWalkInOnboardingMetrics,
+	buildWalkInStatItems,
+	formatShareCardCreatedAt,
+	getWrappedArchetypeIndex,
+} from "@/features/walk-in/walk-in-team-card-models";
 import { useWalkInCardData } from "@/features/walk-in/use-walk-in-card-data";
 import {
 	type WalkInCardTiltController,
@@ -45,10 +43,7 @@ import {
 	type WalkInTeamMemberCardTheme,
 } from "@/features/walk-in/WalkInTeamMemberCard";
 import { MAX_ANALYTICS_DAYS } from "@/lib/analytics-date-range";
-import {
-	formatCompactWholeCurrency,
-	formatCompactWholeNumber,
-} from "@/lib/format";
+import { formatCompactWholeCurrency } from "@/lib/format";
 import { orpc } from "@/lib/orpc";
 import {
 	captureElement,
@@ -762,293 +757,6 @@ function TeamCardWalkInPageContent(props: {
 	);
 }
 
-function buildResolvedTeamCardRow(params: {
-	accountLabel: string;
-	debugProfileImageSrc: string;
-	developerDetails: DeveloperDetails | undefined;
-	sessionUserEmail: string | undefined;
-	sessionUserId: string | undefined;
-	sessionUserName: string | undefined;
-	teamMemberRows: readonly TeamPageMemberRow[];
-}): TeamPageMemberRow {
-	const {
-		accountLabel,
-		debugProfileImageSrc,
-		developerDetails,
-		sessionUserEmail,
-		sessionUserId,
-		sessionUserName,
-		teamMemberRows,
-	} = params;
-	const currentUserRow = findCurrentUserRow({
-		sessionUserEmail,
-		sessionUserId,
-		teamMemberRows,
-	});
-	const { displayName } = resolveTeamCardDisplayName({
-		accountLabel,
-		currentUserRow,
-		sessionUserEmail,
-		sessionUserName,
-	});
-	const email = currentUserRow?.email ?? sessionUserEmail ?? null;
-
-	if (developerDetails && sessionUserId) {
-		return {
-			activeDays: developerDetails.active_days,
-			cost: developerDetails.cost,
-			displayName,
-			email,
-			favoriteModel: developerDetails.favorite_model,
-			hasActivity:
-				developerDetails.total_sessions > 0 ||
-				developerDetails.active_days > 0 ||
-				developerDetails.total_tokens > 0,
-			imageUrl: debugProfileImageSrc,
-			inputTokens: developerDetails.input_tokens,
-			lastActiveDate: developerDetails.last_active_date,
-			outputTokens: developerDetails.output_tokens,
-			role: currentUserRow?.role ?? "Tracked collaborator",
-			totalSessions: developerDetails.total_sessions,
-			totalTokens: developerDetails.total_tokens,
-			userId: sessionUserId,
-		};
-	}
-
-	if (currentUserRow) {
-		return {
-			...currentUserRow,
-			imageUrl: debugProfileImageSrc,
-		};
-	}
-
-	return {
-		activeDays: 0,
-		cost: 0,
-		displayName,
-		email,
-		favoriteModel: null,
-		hasActivity: false,
-		imageUrl: debugProfileImageSrc,
-		inputTokens: 0,
-		lastActiveDate: null,
-		outputTokens: 0,
-		role: "Tracked collaborator",
-		totalSessions: 0,
-		totalTokens: 0,
-		userId: sessionUserId ?? "walk-in-preview",
-	};
-}
-
-function findCurrentUserRow(input: {
-	sessionUserEmail: string | undefined;
-	sessionUserId: string | undefined;
-	teamMemberRows: readonly TeamPageMemberRow[];
-}) {
-	const { sessionUserEmail, sessionUserId, teamMemberRows } = input;
-	const normalizedSessionEmail = normalizeEmail(sessionUserEmail);
-
-	return teamMemberRows.find((row) => {
-		if (sessionUserId && row.userId === sessionUserId) {
-			return true;
-		}
-
-		return (
-			Boolean(normalizedSessionEmail) &&
-			normalizeEmail(row.email) === normalizedSessionEmail
-		);
-	});
-}
-
-function resolveTeamCardDisplayName(input: {
-	accountLabel: string;
-	currentUserRow: TeamPageMemberRow | undefined;
-	sessionUserEmail: string | undefined;
-	sessionUserName: string | undefined;
-}) {
-	const { accountLabel, currentUserRow, sessionUserEmail, sessionUserName } =
-		input;
-	const meaningfulSessionUserName = getMeaningfulDisplayName(sessionUserName);
-
-	if (meaningfulSessionUserName) {
-		return {
-			displayName: meaningfulSessionUserName,
-			source: "session.name",
-		} as const;
-	}
-
-	const emailHandle = getEmailHandle(sessionUserEmail);
-
-	if (emailHandle) {
-		return {
-			displayName: emailHandle,
-			source: "session.emailHandle",
-		} as const;
-	}
-
-	const meaningfulCurrentUserDisplayName = getMeaningfulDisplayName(
-		currentUserRow?.displayName,
-	);
-
-	if (meaningfulCurrentUserDisplayName) {
-		return {
-			displayName: meaningfulCurrentUserDisplayName,
-			source: "teamRow.displayName",
-		} as const;
-	}
-
-	return {
-		displayName: getFallbackTeamMemberDisplayName(accountLabel),
-		source: "accountLabelFallback",
-	} as const;
-}
-
-function buildWalkInStatItems(
-	row: TeamPageMemberRow,
-	distinctProjectCount: number,
-	sourceSplit: readonly WrappedSourceSplit[],
-): WalkInTeamMemberCardStatItem[] {
-	const normalizedSourceSplit = normalizeSourceSplit(sourceSplit);
-
-	return [
-		{
-			key: "codex-share",
-			title: `${normalizedSourceSplit.codexShare}% of wrapped sessions came from Codex`,
-			icon: "codex",
-			value: `${normalizedSourceSplit.codexShare}%`,
-		},
-		{
-			key: "claude-share",
-			title: `${normalizedSourceSplit.claudeShare}% of wrapped sessions came from Claude Code`,
-			icon: "claude",
-			value: `${normalizedSourceSplit.claudeShare}%`,
-		},
-		{
-			key: "sessions",
-			label: "SESS",
-			title: `${row.totalSessions.toLocaleString()} sessions`,
-			value: row.totalSessions.toLocaleString(),
-		},
-		{
-			key: "days",
-			label: "DAYS",
-			title: `${row.activeDays.toLocaleString()} active days across ${MAX_ANALYTICS_DAYS.toLocaleString()} tracked days`,
-			value: row.activeDays.toLocaleString(),
-		},
-		{
-			key: "tokens",
-			label: "TOK",
-			title: `${row.totalTokens.toLocaleString()} total tokens`,
-			value: formatCompactWholeNumber(row.totalTokens),
-		},
-		{
-			key: "repos",
-			label: "REPOS",
-			title: `${distinctProjectCount.toLocaleString()} distinct tracked projects`,
-			value: distinctProjectCount.toLocaleString(),
-		},
-	];
-}
-
-function buildWalkInOnboardingMetrics(input: {
-	commitBreakdown: readonly DimensionAnalysisDataPoint[] | undefined;
-	developerDetails: DeveloperDetails | undefined;
-	developerFeatures: DeveloperFeatureUsage | undefined;
-	developerProjects: readonly DeveloperProject[] | undefined;
-	developerSessions: readonly DeveloperSession[] | undefined;
-	wrappedMetrics: WrappedV1["metrics"] | undefined;
-}): WalkInOnboardingMetrics {
-	const {
-		commitBreakdown,
-		developerDetails,
-		developerFeatures,
-		developerProjects,
-		developerSessions,
-		wrappedMetrics,
-	} = input;
-	const totalSessions = developerDetails?.total_sessions ?? 0;
-	const commitSessions = findBooleanDimensionCount(commitBreakdown, true);
-	const topProject = findTopProject(developerProjects);
-	const repoPulse = buildRepoPulse(developerSessions);
-
-	return {
-		activeDays:
-			wrappedMetrics?.active_days ?? developerDetails?.active_days ?? 0,
-		avgSessionMin: developerDetails?.avg_session_duration_min ?? null,
-		commitRate:
-			totalSessions > 0 ? (commitSessions / totalSessions) * 100 : null,
-		daysSinceFirst: wrappedMetrics?.days_since_first_session ?? 0,
-		favoriteModel: formatWalkInLabel(
-			wrappedMetrics?.favorite_model ??
-				developerDetails?.favorite_model ??
-				undefined,
-		),
-		longestSessionMin: wrappedMetrics?.longest_session_min ?? null,
-		modelByMonth: wrappedMetrics?.model_by_month ?? [],
-		sourceSplit: wrappedMetrics?.source_split ?? [],
-		repoPulse,
-		skillsAdoptionRate: developerFeatures?.skills_adoption_rate ?? null,
-		slashCommandsAdoptionRate:
-			developerFeatures?.slash_commands_adoption_rate ?? null,
-		subagentsAdoptionRate: developerFeatures?.subagents_adoption_rate ?? null,
-		successRate: developerDetails?.success_rate ?? null,
-		topProjectName: getProjectDisplayName(topProject),
-		topProjectSessions: topProject?.sessions ?? 0,
-		topProjectTokens: topProject?.total_tokens ?? 0,
-		topSkills:
-			developerFeatures?.top_skills
-				.map((skill) => ({
-					count: skill.count,
-					name: formatWalkInLabel(skill.name),
-				}))
-				.filter(
-					(
-						skill,
-					): skill is {
-						count: number;
-						name: string;
-					} => Boolean(skill.name),
-				) ?? [],
-		topSlashCommand: formatWalkInLabel(
-			developerFeatures?.top_slash_commands[0]?.name,
-		),
-		topSlashCommands:
-			developerFeatures?.top_slash_commands
-				.map((command) => ({
-					count: command.count,
-					name: formatWalkInLabel(command.name),
-				}))
-				.filter(
-					(
-						command,
-					): command is {
-						count: number;
-						name: string;
-					} => Boolean(command.name),
-				) ?? [],
-		topSlashCommandCount: developerFeatures?.top_slash_commands[0]?.count ?? null,
-		topSubagent: formatWalkInLabel(developerFeatures?.top_subagents[0]?.name),
-		topSubagents:
-			developerFeatures?.top_subagents
-				.map((subagent) => ({
-					count: subagent.count,
-					name: formatWalkInLabel(subagent.name),
-				}))
-				.filter(
-					(
-						subagent,
-					): subagent is {
-						count: number;
-						name: string;
-					} => Boolean(subagent.name),
-				) ?? [],
-		topSubagentCount: developerFeatures?.top_subagents[0]?.count ?? null,
-		totalSessions,
-		totalTokens:
-			wrappedMetrics?.total_tokens ?? developerDetails?.total_tokens ?? 0,
-	};
-}
-
 function getSessionUserId(
 	session: ReturnType<typeof useWalkInCardData>["session"],
 ) {
@@ -1069,44 +777,6 @@ function getActiveMemberUserId(
 		: undefined;
 }
 
-function getFallbackTeamMemberDisplayName(accountLabel: string): string {
-	if (accountLabel.includes("@")) {
-		return accountLabel.split("@")[0] || "User";
-	}
-
-	return accountLabel || "User";
-}
-
-function getEmailHandle(email: string | undefined) {
-	if (!email) {
-		return undefined;
-	}
-
-	const [emailHandle] = email.split("@");
-	return emailHandle?.trim() || undefined;
-}
-
-function normalizeEmail(email: string | null | undefined) {
-	return email?.trim().toLowerCase() || undefined;
-}
-
-function getMeaningfulDisplayName(value: string | undefined) {
-	const normalizedValue = value?.trim();
-
-	if (!normalizedValue) {
-		return undefined;
-	}
-
-	if (
-		normalizedValue.toLowerCase() === "operator" ||
-		normalizedValue.toLowerCase() === "unknown teammate"
-	) {
-		return undefined;
-	}
-
-	return normalizedValue;
-}
-
 function getSessionUserName(
 	session: ReturnType<typeof useWalkInCardData>["session"],
 ) {
@@ -1125,294 +795,4 @@ function getSessionUserEmail(
 		typeof session.user.email === "string"
 		? session.user.email
 		: undefined;
-}
-
-function getWrappedArchetypeIndex(index: number, count: number) {
-	return ((index % count) + count) % count;
-}
-
-function normalizeSourceSplit(sourceSplit: readonly WrappedSourceSplit[]) {
-	return {
-		claudeShare: Math.round(getSourceSharePercent(sourceSplit, "claude_code")),
-		codexShare: Math.round(getSourceSharePercent(sourceSplit, "codex")),
-	};
-}
-
-function formatShareCardCreatedAt(date: Date) {
-	return new Intl.DateTimeFormat("en-US", {
-		month: "2-digit",
-		day: "2-digit",
-		year: "numeric",
-	}).format(date);
-}
-
-function getSourceSharePercent(
-	sourceSplit: readonly WrappedSourceSplit[],
-	source: WrappedSourceSplit["source"],
-) {
-	return (
-		sourceSplit.find((sourceEntry) => sourceEntry.source === source)
-			?.session_share_percent ?? 0
-	);
-}
-
-function formatWalkInLabel(value: string | undefined) {
-	const trimmedValue = value?.trim();
-
-	if (!trimmedValue) {
-		return null;
-	}
-
-	return trimmedValue
-		.replaceAll(/[_-]+/g, " ")
-		.replaceAll(/\s+/g, " ")
-		.trim()
-		.replaceAll(/\b\w/g, (character) => character.toUpperCase());
-}
-
-function getMetricValue(row: DimensionAnalysisDataPoint) {
-	return Number(row.metric_value) || 0;
-}
-
-function findBooleanDimensionCount(
-	rows: readonly DimensionAnalysisDataPoint[] | undefined,
-	target: boolean,
-) {
-	const match = rows?.find((row) => {
-		const normalizedValue = row.dimension_value.trim().toLowerCase();
-		return target
-			? normalizedValue === "true" || normalizedValue === "1"
-			: normalizedValue === "false" || normalizedValue === "0";
-	});
-
-	return match ? getMetricValue(match) : 0;
-}
-
-function findTopProject(projects: readonly DeveloperProject[] | undefined) {
-	return [...(projects ?? [])].sort(
-		(leftRow, rightRow) =>
-			rightRow.total_tokens - leftRow.total_tokens ||
-			rightRow.sessions - leftRow.sessions ||
-			leftRow.project_path.localeCompare(rightRow.project_path),
-	)[0];
-}
-
-function buildRepoPulse(
-	sessions: readonly DeveloperSession[] | undefined,
-): WalkInOnboardingMetrics["repoPulse"] {
-	const repoStats = new Map<
-		string,
-		{
-			errorSessions: number;
-			repoName: string;
-			sessionCount: number;
-			skillSessions: number;
-			slashSessions: number;
-			subagentSessions: number;
-			successSessions: number;
-			totalDurationMin: number;
-			totalTokens: number;
-		}
-	>();
-
-	for (const session of sessions ?? []) {
-		const repoKey = getRepoPulseProjectKey(session);
-		const repoLabel = getProjectDisplayName(session);
-
-		if (!repoLabel) {
-			continue;
-		}
-
-		const existingStats = repoStats.get(repoKey);
-		repoStats.set(repoKey, {
-			errorSessions:
-				(existingStats?.errorSessions ?? 0) + Number(session.has_errors),
-			repoName: repoLabel,
-			sessionCount: (existingStats?.sessionCount ?? 0) + 1,
-			skillSessions:
-				(existingStats?.skillSessions ?? 0) + Number(session.has_skills),
-			slashSessions:
-				(existingStats?.slashSessions ?? 0) + Number(session.has_slash_commands),
-			subagentSessions:
-				(existingStats?.subagentSessions ?? 0) + Number(session.has_subagents),
-			successSessions:
-				(existingStats?.successSessions ?? 0) + Number(session.likely_success),
-			totalDurationMin:
-				(existingStats?.totalDurationMin ?? 0) + session.duration_min,
-			totalTokens: (existingStats?.totalTokens ?? 0) + session.total_tokens,
-		});
-	}
-
-	if (repoStats.size === 0) {
-		return {
-			entries: [],
-			leadRepoName: null,
-			totalRepos: 0,
-			totalSessions: 0,
-		};
-	}
-
-	const rankedRepos = [...repoStats.entries()].sort(
-		(leftEntry, rightEntry) =>
-			rightEntry[1].sessionCount - leftEntry[1].sessionCount ||
-			rightEntry[1].totalTokens - leftEntry[1].totalTokens ||
-			rightEntry[1].totalDurationMin - leftEntry[1].totalDurationMin ||
-			leftEntry[1].repoName.localeCompare(rightEntry[1].repoName),
-	);
-	const entries = rankedRepos.slice(0, 3).map(([repoKey, stats]) => {
-		const workType = resolveRepoPulseWorkType(stats);
-
-		return {
-			id: `repo-pulse-${repoKey}`,
-			meta: buildRepoPulseMeta(stats),
-			proof: workType.proof,
-			repoName: stats.repoName,
-			workType: workType.label,
-		};
-	});
-	const totalSessions = rankedRepos.reduce(
-		(sum, [, stats]) => sum + stats.sessionCount,
-		0,
-	);
-
-	return {
-		entries,
-		leadRepoName: entries[0]?.repoName ?? null,
-		totalRepos: rankedRepos.length,
-		totalSessions,
-	};
-}
-
-function getRepoPulseProjectKey(project: {
-	git_remote?: string;
-	package_name?: string;
-	project_path: string;
-}) {
-	return (
-		project.project_path ??
-		project.git_remote ??
-		getProjectDisplayName(project) ??
-		"unknown-project"
-	);
-}
-
-function getProjectDisplayName(
-	project:
-		| {
-				git_remote?: string;
-				package_name?: string;
-				project_path: string;
-		  }
-		| undefined,
-) {
-	if (!project) {
-		return null;
-	}
-
-	const packageName = project.package_name?.trim();
-
-	if (packageName) {
-		return packageName;
-	}
-
-	const remoteName = project.git_remote?.split("/").pop()?.trim();
-
-	if (remoteName) {
-		return remoteName.replace(/\.git$/i, "");
-	}
-
-	const projectPath = project.project_path?.trim();
-
-	return projectPath || null;
-}
-
-function resolveRepoPulseWorkType(stats: {
-	errorSessions: number;
-	sessionCount: number;
-	skillSessions: number;
-	slashSessions: number;
-	subagentSessions: number;
-	successSessions: number;
-	totalDurationMin: number;
-	totalTokens: number;
-}) {
-	const avgDurationMin = stats.totalDurationMin / stats.sessionCount;
-	const avgTokens = stats.totalTokens / stats.sessionCount;
-	const skillRate = (stats.skillSessions / stats.sessionCount) * 100;
-	const slashRate = (stats.slashSessions / stats.sessionCount) * 100;
-	const subagentRate = (stats.subagentSessions / stats.sessionCount) * 100;
-	const successRate = (stats.successSessions / stats.sessionCount) * 100;
-
-	if (subagentRate >= 25) {
-		return {
-			label: "Delegating",
-			proof: `${Math.round(subagentRate)}% used subagents`,
-		};
-	}
-
-	if (skillRate >= 28) {
-		return {
-			label: "Skills-heavy",
-			proof: `${Math.round(skillRate)}% used skills`,
-		};
-	}
-
-	if (slashRate >= 28) {
-		return {
-			label: "Command-heavy",
-			proof: `${Math.round(slashRate)}% used slash commands`,
-		};
-	}
-
-	if (avgDurationMin >= 45) {
-		return {
-			label: "Deep work",
-			proof: `${formatDurationMinutesShort(avgDurationMin)} avg session`,
-		};
-	}
-
-	if (avgTokens >= 45_000) {
-		return {
-			label: "Heavy lift",
-			proof: `${formatCompactWholeNumber(Math.round(avgTokens))} tokens / session`,
-		};
-	}
-
-	if (successRate >= 78) {
-		return {
-			label: "Shipping lane",
-			proof: `${Math.round(successRate)}% likely successful`,
-		};
-	}
-
-	return {
-		label: avgDurationMin <= 18 ? "Quick passes" : "Steady work",
-		proof:
-			avgDurationMin <= 18
-				? `${formatDurationMinutesShort(avgDurationMin)} avg session`
-				: `${Math.round(successRate)}% likely successful`,
-	};
-}
-
-function buildRepoPulseMeta(stats: {
-	sessionCount: number;
-	totalDurationMin: number;
-	totalTokens: number;
-}) {
-	return `${stats.sessionCount.toLocaleString()} sessions · ${formatDurationMinutesShort(stats.totalDurationMin)} total`;
-}
-
-function formatDurationMinutesShort(totalDurationMin: number) {
-	if (totalDurationMin < 60) {
-		return `${Math.round(totalDurationMin)}m`;
-	}
-
-	const hours = Math.floor(totalDurationMin / 60);
-	const minutes = Math.round(totalDurationMin - hours * 60);
-
-	if (minutes === 0) {
-		return `${hours}h`;
-	}
-
-	return `${hours}h ${minutes}m`;
 }
